rename test packages, extend pom.xml
