+++ src/test/java/domain/BenaTest.java
@@ -1,7 +1,5 @@
-package Junit;
+package domain;
 
-import domain.Bena;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
