+++ src/test/java/domain/LepesInvalidArgumentsProvider.java
@@ -1,4 +1,4 @@
-package Junit;
+package domain;
 
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
