+++ src/test/java/domain/LepesmgrTest.java
@@ -1,7 +1,5 @@
-package Junit;
+package domain;
 
-import domain.*;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
