+++ src/test/java/domain/VitustancoloTest.java
@@ -1,9 +1,5 @@
-package Junit;
+package domain;
 
-import domain.Controller;
-import domain.Vitustancolo;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
