+++ src/test/java/frames/RunCucumberTest.java
@@ -1,4 +1,4 @@
-package cucumber;
+package frames;
 
 //import org.junit.platform.suite.api.ConfigurationParameter;
 //import org.junit.platform.suite.api.IncludeEngines;
