+++ src/test/java/frames/TestCloak.java
@@ -1,4 +1,4 @@
-package cucumber;
+package frames;
 
 import domain.Agens;
 import domain.Kopeny;
