+++ src/test/java/frames/TestRandomMove.java
@@ -1,4 +1,4 @@
-package cucumber;
+package frames;
 
 import domain.Vitustancolo;
