+++ src/test/java/frames/TestVirologist.java
@@ -1,4 +1,4 @@
-package cucumber;
+package frames;
 
 import domain.Agens;
 import domain.Hatas;
